+++ packages/mangrove.js/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Next version
 
+# 0.3.1 (March 2022)
+
+- The new logging has been stunted: It only logs to the console and without timestamps
+  - This is a temporary workaround to issue #220
+
 # 0.2.0 (February 2022)
 
 - New `Market` options: